I'm back! Minor bugs fixes and additions.
- Bugs fixed:
--- GE card null reference bug fixed
--- 4th active threat bug fixed
--- Empty SoE bar bug fixed
--- Double dividends bug fixed
- New Global Event Card added: “New Virus”
- Investment spheres labeled properly

+++ Assets/Scripts/Commands/ModifyPlayerTokensCommand.cs
@@ -24,7 +24,7 @@ public class ModifyPlayerTokensCommand : ICommand
         {
             case EffectType.Add:
                 // Add or subtract tokens
-                newTokens = currentTokens + effect.value;
+                newTokens = currentTokens + (int)effect.value;
                 Debug.Log($"Adding {effect.value} tokens to {player.playerName}. {currentTokens} -> {newTokens}");
                 break;
